Merged Plugins, Patched Bugs, and Started on Economy
I added the world manager plugin into this one and the obby destroyer.
Feel free to edit any of those. I fixed the getChunkOwner(Chunk c)
method because the result was being improperly bit shifted. I made
disbanding a faction now clear all claims. I made Vector2D serializable,
so it can be saved now. Last but not least, I added EconomyMain and
started the process. The prices for the items is now saved in an easy to
use yml file.